+++ docs/search/typedefs_4.js
@@ -2,7 +2,9 @@ var searchData=
 [
   ['registerdeviceasynccall_0',['RegisterDeviceAsyncCall',['../classsensory_1_1service_1_1_o_auth_service.html#a3acc6a3b87d25ddf543e9b98c844ea2d',1,'sensory::service::OAuthService']]],
   ['registerdevicecallbackdata_1',['RegisterDeviceCallbackData',['../classsensory_1_1service_1_1_o_auth_service.html#a722752aea768bc6d507b360555148a97',1,'sensory::service::OAuthService']]],
-  ['renewcredentialasynccall_2',['RenewCredentialAsyncCall',['../classsensory_1_1service_1_1_o_auth_service.html#a5072f0575b57641b8adad7b66875ec7d',1,'sensory::service::OAuthService']]],
-  ['renewcredentialcallbackdata_3',['RenewCredentialCallbackData',['../classsensory_1_1service_1_1_o_auth_service.html#af93d22ff5410f22184f155ecd9d0b0ca',1,'sensory::service::OAuthService']]],
-  ['rpc_4',['RPC',['../structsensory_1_1calldata_1_1_async_reader_writer_call.html#ade978303e1c9f5eb25056e35fb4498fe',1,'sensory::calldata::AsyncReaderWriterCall::RPC()'],['../structsensory_1_1calldata_1_1_async_response_reader_call.html#a05301bd072856d106a1e147d5efc1f54',1,'sensory::calldata::AsyncResponseReaderCall::RPC()']]]
+  ['removeenrollmentsfromgroupasynccall_2',['RemoveEnrollmentsFromGroupAsyncCall',['../classsensory_1_1service_1_1_management_service.html#af4991dec148303d7a747a5d193ba9893',1,'sensory::service::ManagementService']]],
+  ['removeenrollmentsfromgroupcallbackdata_3',['RemoveEnrollmentsFromGroupCallbackData',['../classsensory_1_1service_1_1_management_service.html#a0cf679f0d462ef6faf55042f3ced805c',1,'sensory::service::ManagementService']]],
+  ['renewcredentialasynccall_4',['RenewCredentialAsyncCall',['../classsensory_1_1service_1_1_o_auth_service.html#a5072f0575b57641b8adad7b66875ec7d',1,'sensory::service::OAuthService']]],
+  ['renewcredentialcallbackdata_5',['RenewCredentialCallbackData',['../classsensory_1_1service_1_1_o_auth_service.html#af93d22ff5410f22184f155ecd9d0b0ca',1,'sensory::service::OAuthService']]],
+  ['rpc_6',['RPC',['../structsensory_1_1calldata_1_1_async_reader_writer_call.html#ade978303e1c9f5eb25056e35fb4498fe',1,'sensory::calldata::AsyncReaderWriterCall::RPC()'],['../structsensory_1_1calldata_1_1_async_response_reader_call.html#a05301bd072856d106a1e147d5efc1f54',1,'sensory::calldata::AsyncResponseReaderCall::RPC()']]]
 ];