+++ examples/management/management_async.cpp
@@ -36,7 +36,7 @@ using sensory::service::ManagementService;
 
 /// @brief Get the enrollments for the given user.
 ///
-/// @param service The management service for getting enrollments.
+/// @param service The management service to interact with.
 /// @returns 0 if the call succeeds, 1 otherwise.
 ///
 int get_enrollments(
@@ -100,7 +100,7 @@ int delete_enrollment(
 
 /// @brief Get the enrollment groups for the given user.
 ///
-/// @param service The management service for getting enrollments.
+/// @param service The management service to interact with.
 /// @returns 0 if the call succeeds, 1 otherwise.
 ///
 int get_enrollment_groups(
@@ -138,7 +138,7 @@ int get_enrollment_groups(
 
 /// @brief Create a new enrollment group.
 ///
-/// @param service The management service for getting enrollments.
+/// @param service The management service to interact with.
 /// @param userID The user ID of the user that owns the enrollment group.
 /// @param groupdID The optional group ID to create the group with (an empty
 /// string to generate the UUID on the server).
@@ -168,7 +168,37 @@ int create_enrollment_group(
             std::cout << "Failed to create enrollment group (" << call->getStatus().error_code() << "): " << call->getStatus().error_message() << std::endl;
             error_code = call->getStatus().error_code();
         } else {
-            std::cout << "Created group with ID " << groupID << std::endl;
+            std::cout << "Created group with ID " << call->getResponse().id() << std::endl;
+        }
+        delete call;
+    }
+    return error_code;
+}
+
+/// @brief Update an existing enrollment group.
+///
+/// @param service The management service to interact with.
+/// @param groupdID The ID of the enrollment group to update
+/// @param groupName The new name for the group.
+/// @returns 0 if the call succeeds, 1 otherwise.
+///
+int update_enrollment_group(
+    ManagementService<FileSystemCredentialStore>& service,
+    const std::string& groupID,
+    const std::string& groupName
+) {
+    ::grpc::CompletionQueue queue;
+    auto call = service.update_enrollment_group(&queue, groupID, groupName);
+    void* tag(nullptr);
+    bool ok(false);
+    queue.Next(&tag, &ok);
+    int error_code = 0;
+    if (ok && tag == call) {
+        if (!call->getStatus().ok()) {  // The call failed.
+            std::cout << "Failed to update enrollment group (" << call->getStatus().error_code() << "): " << call->getStatus().error_message() << std::endl;
+            error_code = call->getStatus().error_code();
+        } else {
+            std::cout << "Updated group with ID " << groupID << " to have name " << groupName << std::endl;
         }
         delete call;
     }
@@ -177,7 +207,7 @@ int create_enrollment_group(
 
 /// @brief Append enrollment IDs to an existing enrollment group.
 ///
-/// @param service The management service for getting enrollments.
+/// @param service The management service to interact with.
 /// @param groupID the UUID of the group to append enrollments to
 /// @param enrollments the list of enrollments to append to the group
 /// @returns 0 if the call succeeds, 1 otherwise.
@@ -196,6 +226,34 @@ int append_enrollment_group(
     if (ok && tag == call) {
         if (!call->getStatus().ok()) {  // The call failed.
             std::cout << "Failed to append enrollment group (" << call->getStatus().error_code() << "): " << call->getStatus().error_message() << std::endl;
+            error_code = call->getStatus().error_code();
+        }
+        delete call;
+    }
+    return error_code;
+}
+
+/// @brief Remove enrollments from an existing enrollment group.
+///
+/// @param service The management service to interact with.
+/// @param groupID the UUID of the group to remove enrollments from
+/// @param enrollments the list of enrollments to remove from the group
+/// @returns 0 if the call succeeds, 1 otherwise.
+///
+int remove_enrollments_from_group(
+    ManagementService<FileSystemCredentialStore>& service,
+    const std::string& groupID,
+    const std::vector<std::string>& enrollments
+) {
+    ::grpc::CompletionQueue queue;
+    auto call = service.remove_enrollments_from_group(&queue, groupID, enrollments);
+    void* tag(nullptr);
+    bool ok(false);
+    queue.Next(&tag, &ok);
+    int error_code = 0;
+    if (ok && tag == call) {
+        if (!call->getStatus().ok()) {  // The call failed.
+            std::cout << "Failed to remove enrollments from group (" << call->getStatus().error_code() << "): " << call->getStatus().error_message() << std::endl;
             error_code = call->getStatus().error_code();
         }
         delete call;
@@ -244,6 +302,8 @@ int main(int argc, const char** argv) {
             "get_enrollment_groups",
             "create_enrollment_group",
             "append_enrollment_group",
+            "update_enrollment_group",
+            "remove_enrollments_from_group",
             "delete_enrollment_group"
         }).help("The management endpoint to use.");
     parser.add_argument({ "-u", "--userid" })
@@ -308,10 +368,20 @@ int main(int argc, const char** argv) {
         return 1;
     }
 
-    if      (ENDPOINT == "get_enrollments")        return get_enrollments(cloud.management, USER_ID);
-    else if (ENDPOINT == "delete_enrollment")      return delete_enrollment(cloud.management, ENROLLMENT_ID);
-    else if (ENDPOINT == "get_enrollment_groups")   return get_enrollment_groups(cloud.management, USER_ID);
-    else if (ENDPOINT == "create_enrollment_group") return create_enrollment_group(cloud.management, USER_ID, ENROLLMENT_ID, NAME, DESCRIPTION, MODEL, {});
-    else if (ENDPOINT == "append_enrollment_group") return append_enrollment_group(cloud.management, ENROLLMENT_ID, ENROLLMENT_IDS);
-    else if (ENDPOINT == "delete_enrollment_group") return delete_enrollment_group(cloud.management, ENROLLMENT_ID);
+    if      (ENDPOINT == "get_enrollments")
+        return get_enrollments(cloud.management, USER_ID);
+    else if (ENDPOINT == "delete_enrollment")
+        return delete_enrollment(cloud.management, ENROLLMENT_ID);
+    else if (ENDPOINT == "get_enrollment_groups")
+        return get_enrollment_groups(cloud.management, USER_ID);
+    else if (ENDPOINT == "create_enrollment_group")
+        return create_enrollment_group(cloud.management, USER_ID, ENROLLMENT_ID, NAME, DESCRIPTION, MODEL, ENROLLMENT_IDS);
+    else if (ENDPOINT == "append_enrollment_group")
+        return append_enrollment_group(cloud.management, ENROLLMENT_ID, ENROLLMENT_IDS);
+    else if (ENDPOINT == "update_enrollment_group")
+        return update_enrollment_group(cloud.management, ENROLLMENT_ID, NAME);
+    else if (ENDPOINT == "remove_enrollments_from_group")
+        return remove_enrollments_from_group(cloud.management, ENROLLMENT_ID, ENROLLMENT_IDS);
+    else if (ENDPOINT == "delete_enrollment_group")
+        return delete_enrollment_group(cloud.management, ENROLLMENT_ID);
 }